Worker/src/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Contenedor encargado de publicar endpoints desde la base de datos"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "celery>=5.6.3",
    "google-auth>=2.49.1",
    "google-genai>=1.69.0",
    "uvicorn>=0.42.0",
]
